deviding file
